+++ cmd/main.go
@@ -46,13 +46,30 @@ func main() {
 				return fmt.Errorf("failed to read max-functions flag: %w", err)
 			}
 
-			p := parser.NewParser(f, out, maxFunctions)
+			outputHandlingStr, err := cmd.Flags().GetString("output-multiline")
+			if err != nil {
+				return fmt.Errorf("failed to read output-multiline flag: %w", err)
+			}
+			var outputHandling parser.MultiLineOutputHandling
+			switch outputHandlingStr {
+			case "preserved":
+				outputHandling = parser.MultilinePreserved
+			case "squashed":
+				outputHandling = parser.MultilineToOneLine
+			case "indented":
+				outputHandling = parser.MultilineWithIndent
+			default:
+				return fmt.Errorf("unknown value [%s] for `output-multiline` supplied, supported: [preserved, squashed, indented]", outputHandlingStr)
+			}
+
+			p := parser.NewParser(f, out, parser.WithMaxFunctionCount(maxFunctions), parser.WithMultilineOutputHandling(outputHandling))
 			return p.Parse(cmd.Context())
 		},
 	}
 
 	cmd.Flags().StringP("output-file", "f", "", "path to the file where result will be saved to, if not set, stdOut is used")
 	cmd.Flags().Int("max-functions", 200, "max amount of function calls that may occur during parsing of the provided file")
+	cmd.Flags().StringP("output-multiline", "o", "indented", "Sets how multiline output of functions will be formatted. Options: `preserved`, `squashed`, `indented`")
 
 	if err := cmd.Execute(); err != nil {
 		metaErr := new(metaError.MetaError)