+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.0] - 2024-03-07
+
+### Added
+- support for `OptionFunc` to configure the parser
+- way to customize how multiline output is handled.
+  - It can be either preserved, squashed to one line or preserved with indent (e.g. for yaml)
+
+### Changed
+- parser constructor now takes variadic parameter `OptionFunc`
+
+### Removed
+- `maxFunctionCount` parameter from constructor
+
+### Fixed
+- issue where multiline function output would not be indented if it was nested inside of text block.
+
 ## [v1.2.2] - 2024-02-28
 
 ### Changed